+++ src/main/java/com/georgster/util/permissions/PermissibleAction.java
@@ -5,7 +5,8 @@ package com.georgster.util.permissions;
  * Any action not declared by a Command will be considered a {@link #DEFAULT} action.
  */
 public enum PermissibleAction {
-    MANAGEEVENT,
+    MANAGEEVENTS,
+    MANAGEALLEVENTS,
     MANAGEPERMISSIONS,
     ADMIN,
     HELPCOMMAND,